atd-etl/afd_ems_import/requirements.txt
onepasswordconnectsdk==1.*
psycopg2-binary==2.*
boto3==1.*
pandas==2.*
sshtunnel==0.*